requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
etelemetry==0.3.0
exceptiongroup==1.1.0
filelock==3.9.0
fitz==0.0.1.dev2
future==0.18.3
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
looseversion==1.0.3
lxml==4.9.2
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
pdf2image==1.16.2
Pillow==9.4.0
pluggy==1.0.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.21.1
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
pyxnat==1.5
rdflib==6.2.0
requests==2.28.2
scipy==1.10.0
simplejson==3.18.1
six==1.16.0
tomli==2.0.1
traits==6.3.2
urllib3==1.26.14
